Added local edge state methods.
