+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/edge/PojoExEdge.scala
@@ -1,7 +1,6 @@
 package com.raphtory.internals.storage.pojograph.entities.external.edge
 
-import com.raphtory.api.analysis.visitor.IndexedValue
-import com.raphtory.api.analysis.visitor.TimePoint
+import com.raphtory.api.analysis.visitor.{HistoricEvent, IndexedValue, TimePoint}
 import com.raphtory.internals.storage.pojograph.PojoGraphLens
 import com.raphtory.internals.storage.pojograph.entities.external.PojoExEntity
 import com.raphtory.internals.storage.pojograph.entities.internal.PojoEdge
@@ -30,55 +29,6 @@ private[pojograph] class PojoExEdge(
 
   def dst: Long = edge.getDstId
 
-  private var computationValues: Map[String, Any] =
-    Map.empty //Partial results kept between supersteps in calculation
-
-  def setState(key: String, value: Any): Unit =
-    computationValues += ((key, value))
-
-  def getState[T](key: String, includeProperties: Boolean = false): T =
-    if (computationValues.contains(key))
-      computationValues(key).asInstanceOf[T]
-    else if (includeProperties && properties.contains(key))
-      getProperty[T](key).get
-    else if (includeProperties)
-      throw new Exception(
-        s"$key not found within analytical state or properties for edge ${(src,dst)}"
-      )
-    else
-      throw new Exception(s"$key not found within analytical state for edge ${(src,dst)}")
-
-  def getStateOrElse[T](key: String, value: T, includeProperties: Boolean = false): T =
-    if (computationValues contains key)
-      computationValues(key).asInstanceOf[T]
-    else if (includeProperties && properties.contains(key))
-      getProperty[T](key).get
-    else
-      value
-
-  def containsState(key: String, includeProperties: Boolean = false): Boolean =
-    computationValues.contains(key) || (includeProperties && properties.contains(key))
-
-  def getOrSetState[T](key: String, value: T, includeProperties: Boolean = false): T = {
-    var output_value = value
-    if (containsState(key))
-      output_value = getState[T](key)
-    else {
-      if (includeProperties && properties.contains(key))
-        output_value = getProperty[T](key).get
-      setState(key, output_value)
-    }
-    output_value
-  }
-
-  def appendToState[T: ClassTag](key: String, value: T): Unit = //write function later
-    computationValues.get(key) match {
-      case Some(arr) =>
-        setState(key, arr.asInstanceOf[Array[T]] :+ value)
-      case None      =>
-        setState(key, Array(value))
-    }
-
   override def explode(): List[ExplodedEdge]                 =
     historyView.collect { case event if event.event => PojoExplodedEdge.fromEdge(this, event) }.toList
 
@@ -93,6 +43,8 @@ private[pojograph] class PojoExEdge(
       new PojoExInOutEdge(this, other, asInEdge)
     else
       new PojoExInOutEdge(other, this, asInEdge)
+
+  override def history(): List[HistoricEvent] = ???
 }
 
 private[pojograph] class PojoExInOutEdge(
@@ -168,54 +120,6 @@ private[pojograph] class PojoExInOutEdge(
     }
     mergedIter.toList
   }
-
-  override def setState(key: String, value: Any): Unit = ???
-
-  /** Retrieve value from algorithmic state
-    *
-    * @tparam `T` value type for state
-    * @param key               key to use for retrieving state
-    * @param includeProperties set this to `true` to fall-through to vertex properties if `key` is not found
-    */
-  override def getState[T](key: String, includeProperties: Boolean): T = ???
-
-  /** Retrieve value from algorithmic state if it exists or return a default value otherwise
-    *
-    * @tparam `T` value type for state
-    * @param key               key to use for retrieving state
-    * @param value             default value to return if state does not exist
-    * @param includeProperties set this to `true` to fall-through to vertex properties
-    *                          if `key` is not found in algorithmic state
-    */
-override def getStateOrElse[T](key: String, value: T, includeProperties: Boolean): T = ???
-
-  /** Checks if algorithmic state with key `key` exists
-    *
-    * @param key               state key to check
-    * @param includeProperties Set this to `true` to fall-through to vertex properties if `key` is not found.
-    *                          If set, this function only returns `false` if `key` is not included in either algorithmic state
-    *                          or vertex properties
-    */
-  override def containsState(key: String, includeProperties: Boolean): Boolean = ???
-
-  /** Retrieve value from algorithmic state if it exists or set this state to a default value and return otherwise
-    *
-    * @tparam `T` value type for state
-    * @param key               key to use for retrieving state
-    * @param value             default value to set and return if state does not exist
-    * @param includeProperties set this to `true` to fall-through to vertex properties
-    *                          if `key` is not found in algorithmic state. State is only set if this is also not found.
-    */
-  override def getOrSetState[T](key: String, value: T, includeProperties: Boolean): T = ???
-
-  /** Append new value to existing array or initialise new array if state does not exist
-    * The value type of the state is assumed to be `Array[T]` if the state already exists.
-    *
-    * @tparam `T` value type for state (needs to have a `ClassTag` available due to Scala `Array` implementation)
-    * @param key   key to use for retrieving state
-    * @param value value to append to state
-    */
-  override def appendToState[T: ClassTag](key: String, value: T): Unit = ???
 }
 
 private[pojograph] class PojoExReversedEdge(
